requirements.txt
azure-functions==1.17.0
azure-functions-worker==1.0.0
psutil==5.9.8
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pg8000==1.30.5 